oldrequirements.txt
atomicwrites==1.4.0
attrs==20.3.0
click==8.0.1
colorama==0.4.4
Flask==2.0.1
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==3.7.2
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
mariadb==1.0.8
MarkupSafe==2.0.1
mysql-connector-python==8.0.23
packaging==20.9
pluggy==0.13.1
protobuf==3.15.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
setuptools==41.2.0
six==1.15.0
SQLAlchemy==1.4.18
sqlalchemy-datatables==2.0.1
SQLAlchemy-serializer==1.4.1
toml==0.10.2
typing-extensions==3.7.4.3
Werkzeug==2.0.1
zipp==3.4.1
flask-login==0.5.0
